requirements.txt
Babel
MouseInfo
PyAutoGUI
PyGetWindow
PyMsgBox
PyRect
PyScreeze
aiogram
aiohttp
python-dotenv
aiosignal
async-timeout
attrs
certifi
charset-normalizer
frozenlist
idna
magic-filter
multidict
pip
pyperclip
pytweening
pytz
setuptools
wheel
yarl